In spatial_data_prep file, a methodology to choose between a custom file, wdpa url or no reserve areas has been included. This affected also the naming in the config file.
In data_exploration, the codes for the land data include now all categories even tho they are not present in the specific land area.

legends file include the categories for the custom file passed, which is not used currently, Openeo is preferred.

changes in the Atlite_custom_region are related to the different output in the cells.